+++ src/dev/theskidster/xjge/graphics/Graphics.java
@@ -3,6 +3,7 @@ package dev.theskidster.xjge.graphics;
 import java.nio.FloatBuffer;
 import java.nio.IntBuffer;
 import org.joml.Matrix4f;
+import org.lwjgl.assimp.AIMatrix4x4;
 import static org.lwjgl.opengl.GL30.*;
 
 /**
@@ -62,6 +63,32 @@ public class Graphics {
         glDeleteVertexArrays(vao);
         glDeleteBuffers(vbo);
         glDeleteBuffers(ibo);
+    }
+    
+    static Matrix4f convertFromAssimp(AIMatrix4x4 aiMatrix) {
+        Matrix4f conversion = new Matrix4f();
+        
+        conversion.m00(aiMatrix.a1());
+        conversion.m10(aiMatrix.a2());
+        conversion.m20(aiMatrix.a3());
+        conversion.m30(aiMatrix.a4());
+        
+        conversion.m01(aiMatrix.b1());
+        conversion.m11(aiMatrix.b2());
+        conversion.m21(aiMatrix.b3());
+        conversion.m31(aiMatrix.b4());
+        
+        conversion.m02(aiMatrix.c1());
+        conversion.m12(aiMatrix.c2());
+        conversion.m22(aiMatrix.c3());
+        conversion.m32(aiMatrix.c4());
+        
+        conversion.m03(aiMatrix.d1());
+        conversion.m13(aiMatrix.d2());
+        conversion.m23(aiMatrix.d3());
+        conversion.m33(aiMatrix.d4());
+        
+        return conversion;
     }
     
 }