+++ src/dev/theskidster/xjge/main/Window.java
@@ -7,15 +7,12 @@ import static org.lwjgl.openal.AL10.AL_PAUSED;
 import org.lwjgl.system.MemoryStack;
 import static org.lwjgl.system.MemoryUtil.NULL;
 import static dev.theskidster.xjge.audio.Audio.ALL_SOURCES;
-import dev.theskidster.xjge.graphics.Light;
 import static dev.theskidster.xjge.hardware.InputDevice.*;
 import dev.theskidster.xjge.puppet.discon.DisCon;
 import dev.theskidster.xjge.puppets.Puppets;
-import dev.theskidster.xjge.util.Color;
 import dev.theskidster.xjge.util.LogLevel;
 import dev.theskidster.xjge.util.Logger;
 import dev.theskidster.xjge.util.ServiceLocator;
-import org.joml.Vector3f;
 
 /**
  * @author J Hoffman
@@ -132,16 +129,6 @@ final class Window {
                 if(key == GLFW_KEY_D) Puppets.FREECAM.pressed[3] = (action != GLFW_RELEASE);
                 
                 Puppets.FREECAM.setSpeedBoostEnabled(mods == GLFW_MOD_SHIFT);
-            }
-            
-            if(key == GLFW_KEY_1 && action == GLFW_PRESS) {
-                float x = (float) Math.random() * 40;
-                float y = (float) Math.random() * 40;
-                float z = (float) -(Math.random() * 240);
-                
-                Color color = Color.YELLOW;
-                
-                Game.addLightSource(new Light(0.5f, 0.5f, new Vector3f(x, y, z), color, color));
             }
         });