+++ changelog.md
@@ -3,6 +3,31 @@ All notable changes to this project will be documented in this file. The format 
 
 
 
+## [1.2.0] - 2020-04-29
+
+### Added
+
+- Support for 3D model loading with some basic functions to rotate, translate, and scale objects.
+- New colors "SILVER", "SLATE_GRAY", and "PERIWINKLE" to the default color palette.
+- New "showLightSources" command to terminal that toggles the visibility of light source location icons at runtime.
+- Basic lighting utilities provided through the Level class.
+- New static addLightSource() method to Game class.
+- Support for simple non-interpolated 3D skeletal animation.
+- New test entities; EntityTeapot, Entity3DAnimTest and Entity2DAnimTest- these will likely be replaced/modified in the future.
+- New SpriteAnimation class that facilitates 2D animation.
+
+### Changed
+
+- Changed the field names in the SpriteSheet class to be more meaningful.
+- Minor changes made to the documentation to make it more cohesive.
+- Changed the name of the "model" Matrix4f field in the Graphics class to "modelMatrix" so it wont get confused with instances of the Model class.
+- Fixed the bug where the engine would crash if the terminal was opened after an entity had been destroyed.
+- Logger output with WARNING level severity will now provide stack traces to the console at runtime if an exception has been caught.
+- Minor changes made to the Texture parsing code to help pinpoint what part of the parsing process failed.
+- The SpriteSheet class now provides a default mapping of its texture coordinates through the imgOffsets field.
+
+
+
 ## [1.0.9] - 2020-03-09
 
 ### Added
@@ -73,11 +98,7 @@ All notable changes to this project will be documented in this file. The format 
 - Game state traversal.
 - Logger that tracks significant events occurring within the engine and writes the output to a .txt file.
 
-
-<<<<<<< HEAD
-
-=======
->>>>>>> 9c9b72c0df852387adb2be7874cc82f179269268
+[1.0.9]: https://github.com/theskidster/XJGE/commit/90211a1e8cdc884334377beaf80c7fbf1671965a
 [1.0.5]: https://github.com/theskidster/XJGE/commit/0f82e35a60f807098d05be3c5e02283ff424e0f3
 [1.0.4]: https://github.com/theskidster/XJGE/commit/4bb6e591d3b40f52da837ec27a66d92f4c6e1dbf
 [1.0.0]: https://github.com/theskidster/XJGE/commit/67b3a12dab536e1db056f03fba46988b25752591